Moved Backend class to a new file
